# Conflicts:
#	src/app/login/index.tsx
#	src/store/modules/tool.ts

+++ src/app/setting/index.tsx
@@ -8,6 +8,7 @@ import {useTranslation} from "react-i18next";
 import {Avatar, Button, Dialog, List, Space, Tag} from "antd-mobile";
 import {selectGroupMap} from "@/store/modules/base.ts";
 import {userSignOut} from "@/utils/request/identity.ts";
+import Cookie from "js-cookie";
 
 export default function Setting() {
   const navigate = useNavigate()
@@ -19,19 +20,30 @@ export default function Setting() {
   const dispatch = useDispatch()
 
   const branchCodes = useMemo(() => {
+    if(!identity) return []
     return (identity?.branchIds as string[])?.map(branchId => {
       const branch = groupMap.get(branchId)
-      return branch.branchCode
+      return branch ? branch.branchCode : branchId
     })
   },[identity?.branchIds,groupMap])
 
   const logout = async () => {
-    const response = await userSignOut()
-    if(response.succeeded){
-      dispatch(removeLogin())
-    }
-  }
+    Dialog.confirm({
+      content: t('common.exitAccountTip'),
+      onConfirm: async () => {
+        const response = await userSignOut()
+        if(response.succeeded){
+          Cookie.remove('token')
 
+          localStorage.removeItem('identity')
+          dispatch(removeLogin())
+          setTimeout(() => {
+            navigate('/login')
+          },200)
+        }
+      },
+    })
+  }
 
   return (
     <section className={'containerMain'}>
@@ -69,13 +81,38 @@ export default function Setting() {
         <div className={'mb-5'}>
           <h3 className={'mb-2 font-medium text-(--text)'}>{t('common.routerFinance')}</h3>
           <List mode={'card'} className={'m-auto!'}>
-            <List.Item prefix={<i className={'iconfont icon-chongzhi text-(--text)! text-[1.6rem]!'} />} onClick={() => {navigate('/group/rechargePayment')}}>{t('group.agentRecharge')}</List.Item>
+            <List.Item prefix={<i className={'iconfont icon-chongzhi text-(--text)! text-[1.6rem]!'} />} onClick={() => {navigate('/finance/rechargePayment')}}>{t('group.agentRecharge')}</List.Item>
+            <List.Item prefix={<i className={'iconfont icon-chongzhi text-(--text)! text-[1.6rem]!'} />} onClick={() => {navigate('/finance/rechargePaymentOnline')}}>{t('common.routerRechargePaymentOnline')}</List.Item>
+          </List>
+        </div>
+
+        <div className={'mb-5'}>
+          <h3 className={'mb-2 font-medium text-(--text)'}>{t('common.routerProcurement')}</h3>
+          <List mode={'card'} className={'m-auto!'}>
+            <List.Item prefix={<i className={'iconfont icon-caigoudan text-(--text)! text-[1.6rem]!'} />} onClick={() => {navigate('/foundation/booking')}}>{t('common.routerBookingAccountConfiguration')}</List.Item>
+            <List.Item prefix={<i className={'iconfont icon-caigoudan text-(--text)! text-[1.6rem]!'} />} onClick={() => {navigate('/foundation/payment')}}>{t('common.routerPaymentAccountConfiguration')}</List.Item>
+          </List>
+        </div>
+        <div className={'mb-5'}>
+          <h3 className={'mb-2 font-medium text-(--text)'}>{t('common.routerChannelManagement')}</h3>
+          <List mode={'card'} className={'m-auto!'}>
+            <List.Item prefix={<i className={'iconfont icon-duozhongzhifu text-(--text)! text-[1.6rem]!'} />} onClick={() => {navigate('/channel/payment')}}>{t('common.routerChannelPayment')}</List.Item>
+            <List.Item prefix={<i className={'iconfont icon-caigoudan text-(--text)! text-[1.6rem]!'} />} onClick={() => {navigate('/channel/list')}}>{t('common.routerChannelList')}</List.Item>
+            <List.Item prefix={<i className={'iconfont icon-chongzhi text-(--text)! text-[1.6rem]!'} />} onClick={() => {navigate('/channel/balance')}}>{t('common.routerChannelBalance')}</List.Item>
+          </List>
+        </div>
+        <div className={'mb-5'}>
+          <h3 className={'mb-2 font-medium text-(--text)'}>{t('common.routerBaseData')}</h3>
+          <List mode={'card'} className={'m-auto!'}>
+            <List.Item prefix={<i className={'iconfont icon-plane-trip-international text-(--text)! text-[1.6rem]!'} />} onClick={() => {navigate('/base/nation')}}>{t('common.routerNation')}</List.Item>
+            <List.Item prefix={<i className={'iconfont icon-jichang text-(--text)! text-[1.6rem]!'} />} onClick={() => {navigate('/base/airport')}}>{t('common.routerAirport')}</List.Item>
+            <List.Item prefix={<i className={'iconfont icon-hangdianfeihang text-(--text)! text-[1.6rem]!'} />} onClick={() => {navigate('/base/waypoints')}}>{t('common.routerWaypoints')}</List.Item>
+            <List.Item prefix={<i className={'iconfont icon-huilv01 text-(--text)! text-[1.6rem]!'} />} onClick={() => {navigate('/base/exchangeRate')}}>{t('common.routerExrate')}</List.Item>
           </List>
         </div>
         <div className={'mb-5'}>
           <h3 className={'mb-2 font-medium text-(--text)'}>{t('group.basicConfiguration')}</h3>
           <List mode={'card'} className={'m-auto!'}>
-            <List.Item prefix={<i className={'iconfont icon-caigoudan text-(--text)! text-[1.6rem]!'} />} onClick={() => {navigate('/group/foundation/booking')}}>{t('order.purchasingAccount')}</List.Item>
             <List.Item prefix={<i className={'iconfont icon-yuyan2 text-(--text)! text-[1.6rem]!'} />} onClick={() => {
               Dialog.show({
                 content: t('common.exitAccount'),
